vulnerable-app/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.20
